+++ extensions/one-time-password/CHANGELOG.md
@@ -1,5 +1,11 @@
 # One Time Password Changelog
 
+## [Paste Recent Otp] - {PR_MERGE_DATE}
+
+### Added
+
+- Added "Paste Recent Otp" no-view command that pastes the OTP from the most recently used account into the active app
+
 ## [Windows Support] - 2025-08-27
 
 - Added Windows support.